fixed circular refrenceing and updating model refrencing

+++ src/main/java/com/irby/jaden/resturantreview/domain/core/resturant/repo/ResturantRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResturantRepo extends JpaRepository<Restaurant, Integer> {
+public interface ResturantRepo extends JpaRepository<Restaurant, Long> {
 }
