+++ src/main/java/com/irby/jaden/resturantreview/domain/core/review/repo/ReviewRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReviewRepo extends JpaRepository<Review, Integer> {
+public interface ReviewRepo extends JpaRepository<Review, Long> {
 }
